Tested the Client and Server in the Main file It wowked :3
