+++ .idea/workspace.xml
@@ -4,10 +4,11 @@
     <option name="autoReloadType" value="SELECTIVE" />
   </component>
   <component name="ChangeListManager">
-    <list default="true" id="0cc95419-82d4-4cc3-b2ee-9d1776b196aa" name="Changes" comment="initial commit of the Client and Server Branch&#10;(This branch is dedicated to developing the server and client side of the program)">
-      <change afterPath="$PROJECT_DIR$/src/main/java/org/example/NetWork/Client.java" afterDir="false" />
-      <change afterPath="$PROJECT_DIR$/src/main/java/org/example/NetWork/Server.java" afterDir="false" />
+    <list default="true" id="0cc95419-82d4-4cc3-b2ee-9d1776b196aa" name="Changes" comment="Implemented the Server and Client Class.&#10;(for the time being the Client can only send a message which the server prints, this is just a simple test and will be changed)">
       <change beforePath="$PROJECT_DIR$/.idea/workspace.xml" beforeDir="false" afterPath="$PROJECT_DIR$/.idea/workspace.xml" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/main/java/org/example/Main.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/java/org/example/Main.java" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/main/java/org/example/NetWork/Client.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/java/org/example/NetWork/Client.java" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/main/java/org/example/NetWork/Server.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/java/org/example/NetWork/Server.java" afterDir="false" />
     </list>
     <option name="SHOW_DIALOG" value="false" />
     <option name="HIGHLIGHT_CONFLICTS" value="true" />
@@ -159,7 +160,15 @@
       <option name="project" value="LOCAL" />
       <updated>1718951856195</updated>
     </task>
-    <option name="localTasksCounter" value="12" />
+    <task id="LOCAL-00012" summary="Implemented the Server and Client Class.&#10;(for the time being the Client can only send a message which the server prints, this is just a simple test and will be changed)">
+      <option name="closed" value="true" />
+      <created>1718952609995</created>
+      <option name="number" value="00012" />
+      <option name="presentableId" value="LOCAL-00012" />
+      <option name="project" value="LOCAL" />
+      <updated>1718952609995</updated>
+    </task>
+    <option name="localTasksCounter" value="13" />
     <servers />
   </component>
   <component name="VcsManagerConfiguration">
@@ -175,6 +184,7 @@
     <MESSAGE value="Configured a type for the video file in the Video class." />
     <MESSAGE value="Removed the Blob videFile variable form the Video class &#10;(we'll just save the video locally fuck this shit)" />
     <MESSAGE value="initial commit of the Client and Server Branch&#10;(This branch is dedicated to developing the server and client side of the program)" />
-    <option name="LAST_COMMIT_MESSAGE" value="initial commit of the Client and Server Branch&#10;(This branch is dedicated to developing the server and client side of the program)" />
+    <MESSAGE value="Implemented the Server and Client Class.&#10;(for the time being the Client can only send a message which the server prints, this is just a simple test and will be changed)" />
+    <option name="LAST_COMMIT_MESSAGE" value="Implemented the Server and Client Class.&#10;(for the time being the Client can only send a message which the server prints, this is just a simple test and will be changed)" />
   </component>
 </project>
