+++ src/main/java/org/example/NetWork/Client.java
@@ -6,18 +6,8 @@ import java.net.*;
 
 public class Client
 {
-    void writeMessage(String msg) throws IOException
+    public void writeMessage(String msg) throws IOException
     {
         new ObjectOutputStream((new Socket("localhost",Server.port).getOutputStream())).writeObject(msg);
-    }
-
-    public static void main(String[] args) throws IOException
-    {
-        Server server = new Server();
-        new Thread(server).start();
-        Client client1 = new Client();
-        Client client2 = new Client();
-        client1.writeMessage("Hello !");
-        client2.writeMessage("Give me five !");
     }
 }
